Moved theme into its own file
